backend-py/requirements.txt
fastapi
uvicorn
redis
celery
python-dotenv
groq
python-multipart
requests
